Put in some of the IOP undocumented memory accesses.. the PC now enters the main memory for execution!

+++ PCSX2_Core/Source/Common/Global/GlobalConstants.h
@@ -13,9 +13,9 @@ VALUE = Constant value.
 struct Constants
 {
 	static constexpr size_t SIZE_4GB   = 0x100000000;
-	static constexpr size_t SIZE_512MB = 0x20000000;
-	static constexpr size_t SIZE_4MB   = 0x400000;
-	static constexpr size_t SIZE_16B   = 0x10;
+	static constexpr u32    SIZE_512MB = 0x20000000;
+	static constexpr u32    SIZE_4MB   = 0x400000;
+	static constexpr u32    SIZE_16B   = 0x10;
 
 	static constexpr u32 MASK_16KB = 0x00003FFF;
